% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraq1way.R
\name{fraq1way}
\alias{fraq1way}
\title{fraq1way}
\usage{
fraq1way(
  counts,
  hypothprob,
  conf.level = 0.95,
  addCIs = TRUE,
  digits = 4,
  arrowwid = 0.1,
  estimated = 0
)
}
\arguments{
\item{counts}{num}

\item{hypothprob}{probs}

\item{conf.level}{alpha}

\item{addCIs}{T/F}

\item{digits}{round}

\item{arrowwid}{yes}

\item{estimated}{you got it}
}
\value{
a named list
}
\description{
CI for Differences
}
\examples{
\dontrun{fraq1way(imp)}
}
